fix(security): move ToolBroker to utils so harness keeps I6 isolation
